docs: refactor layout / add stakeholders
This build on top of #445 and restructures how we layout strategic
initiatives. By breaking it out of a table we will be able to include
much more details.

One of the added details here is the inclusion of stakeholders,
individuals outside of the champion who are actively involved in
an initiative. I have only added these for `Modules` at this time
but we should like expand this before landing.

The hope is that this will improve communication and discoverability
as well as offer a new opportunity to recognize significant work.

Refs: https://github.com/nodejs/TSC/issues/442