remove "Core Projects" and "Maintainer" from charter (#421)
* remove core projects from charter

Core Projects are not defined anywhere in the Charter. As far as I know,
we do not have Core Project representation as defined in the Charter. We
do not intend to expand the number of Core Projects, and I'm not even
sure we have any at the current time. Remove this material from the
Charter for clarity and concision. (This will require board approval.
Technically doesn't require TSC approval, but that would be nice.)

* remove maintainer from charter