remove "top level projects"
Top Level Projects feature prominently in the README but are not
actually a thing. There are currently zero Top Level Projects (TLPs)
excluding core. Seems like the usual path nowadays is for a project to
post in the admin repo to ask to be brought under the foundation.

So...simplify our README and remove gaps in information by removing
TLP text.