deleting link to Google doc and announcement
Feel free to revert, but I think that we shouldn't link the Google doc in the published minutes anymore, and I'm also removing the Moderation Team announcement as we usually leave that in the issue tracker and I'm not sure how much value there is in putting the (intentionally vague) information in the minutes. Again, feel free to revert this one. My opinion only.