add strategic initiative for Build resources
From https://github.com/nodejs/build/issues/1154#issuecomment-448418977:

---
Rich and I (Rod) will begin a strategic initiative sometime in January
to do some accounting of resources and costs (people and other) and try
and come up with proposals and plans for reducing pain and improving the
experience for collaborators while managing quality. Or something like
that. Basically we need to think more strategically about this and
understand what our problems and risks are and what we can do about
them.
---

This is that initiative (in March rather than January, but such is
life).