decharter website working group
The Website working group has not been a functioning working group for
some time. (No meetings, charter is frequently iignored, etc.) It is
effectively a GitHub team and I would like to be able to add and remove
people as appropriate, but I can't do that with a working group. Let's
switch it to a GitHub team. We can always recharter at a later date if
necessary.