Update charter to concentrate powers to the TSC rather than its voting members
The only effect of this change should be that TSC regular members participate in lazy consensus process. If you see another effect that I overlooked, that's not on purpose, so please chime in.

The aim of this change is to clarify that TSC voting members do not have any specific powers (except the ones defined by the Moderation Policy and the ones delegated by the TSC on a case by case basis) over Collaborators.